requirements.txt
beautifulsoup4==4.13.3
faiss-cpu==1.10.0
fastapi==0.115.12
mail-parser==4.1.4
numpy==2.2.4
pymupdf==1.25.5
python-docx==1.1.2
rank-bm25==0.2.2
requests==2.32.3
sentence-transformers==3.4.1
streamlit==1.43.2
uvicorn[standard]==0.34.0